COC-164-feat(#75): useError 커스텀 훅 분리
